Entrega1/requirements.txt
mysql-connector
requests
pandas
redis